Refactor globe for better memory managment and ram usage
