+++ src/components/layout/page-container.module.css
@@ -5,6 +5,6 @@
 
 .fullHeight {
   height: 100vh;
-  margin-top: -4.1rem;
-  padding-top: 4.1rem;
+  margin-top: -4.05rem;
+  padding-top: 4.05rem;
 }
