+++ src/components/ui/globe.tsx
@@ -271,7 +271,18 @@ export function World(props: WorldProps) {
   const scene = new Scene();
   scene.fog = new Fog(0xffffff, 400, 2000);
   return (
-    <Canvas scene={scene} camera={new PerspectiveCamera(50, aspect, 180, 1800)}>
+    <Canvas
+      scene={scene}
+      camera={{
+        fov: 50,
+        near: 180,
+        far: 1800,
+        position: [0, 0, cameraZ],
+      }}
+      style={{ width: "100%", height: "100%" }}
+      gl={{ antialias: true, alpha: true }}
+      resize={{ scroll: false, debounce: { scroll: 50, resize: 0 } }}
+    >
       <WebGLRendererConfig />
       <ambientLight color={globeConfig.ambientLight} intensity={0.6} />
       <directionalLight
